This should fix argument shadowing warnings dnk777 reported

+++ source/game/ai/ai_caching_game_allocator.h
@@ -38,8 +38,8 @@ class CachingGameBufferAllocator : UntypedCachingGameAllocator
 	}
 
 public:
-	CachingGameBufferAllocator( const char *tag, size_t limit = 32, int initialCacheSize = 8 )
-		: UntypedCachingGameAllocator( alignedElemSize() * N, tag, limit, (unsigned)initialCacheSize ) {}
+	CachingGameBufferAllocator( const char *tag, size_t limit = 32, unsigned initialCacheSize = 8 )
+		: UntypedCachingGameAllocator( alignedElemSize() * N, tag, limit, initialCacheSize ) {}
 
 	inline T *Alloc() {
 		return (T*)UntypedCachingGameAllocator::Alloc();
